+++ editor.c
@@ -24,7 +24,7 @@ void move_up(Buffer *b) {
 
 void move_right(Buffer *b) {
     char *lp = line_goto(b->data, b->line);
-    size_t len = line_len(lp);
+    uint32_t len = line_len(lp);
     if (b->col < len) {
         b->col++;
         b->col_max = b->col;
@@ -99,7 +99,7 @@ void move_backward(Buffer *b) {
 
 void move_line_begin(Buffer *b) {
     char *cur_line = line_goto(b->data, b->line);
-    size_t len = line_len(cur_line);
+    uint32_t len = line_len(cur_line);
 
     for (size_t i = 0; i < len; ++i) {
         if (cur_line[i] != ' ') {
@@ -111,12 +111,12 @@ void move_line_begin(Buffer *b) {
 
 void move_line_end(Buffer *b) {
     char *cur_line = line_goto(b->data, b->line);
-    size_t len = line_len(cur_line);
+    uint32_t len = line_len(cur_line);
     b->col = b->col_max = len;
 }
 
 /* TODO rename */
-void move_screen_down(Buffer *b, unsigned short int screen_height) {
+void move_screen_down(Buffer *b, uint16_t screen_height) {
     uint16_t n = screen_height / 2;
     uint32_t pos = b->line + n;
     b->line = (pos > line_count(b->data) - 1) ? line_count(b->data) - 1: pos;
@@ -124,16 +124,16 @@ void move_screen_down(Buffer *b, unsigned short int screen_height) {
 }
 
 /* TODO rename */
-void move_screen_up(Buffer *b, size_t screen_height) {
+void move_screen_up(Buffer *b, uint16_t screen_height) {
     uint16_t n = screen_height / 2;
-    int pos = b->line - n;
+    int32_t pos = b->line - n;
     b->line = (pos >= 0) ? pos : 0;
     adjust_col(b);
 }
 
 /* TODO rename */
-void move_screen_center(Buffer *b, size_t screen_height) {
-    int off = b->line - (screen_height / 2);
+void move_screen_center(Buffer *b, uint16_t screen_height) {
+    int32_t off = b->line - (screen_height / 2);
     b->line_off = (off >= 0) ? off : 0;
     adjust_col(b);
 }
@@ -169,7 +169,7 @@ void insert_char(Buffer *b, char c) {
 
 void delete_char(Buffer *b) {
     size_t pos = 0;
-    size_t prev_line_len = 0;
+    uint32_t prev_line_len = 0;
 
     if ((pos = get_current_pos(b)) == 0) return;
     pos--;
@@ -200,9 +200,9 @@ void clear_killed(void) {
 
 void kill_line(Buffer *b) {
     char *cur_line = line_goto(b->data, b->line);
-    size_t size = line_size(cur_line);
+    uint32_t size = line_size(cur_line);
     size_t pos = cur_line - b->data;
-    size_t killed = (cur_line[size] != '\0') ? size + 1 : size;
+    uint32_t killed = (cur_line[size] != '\0') ? size + 1 : size;
 
     strncat(kill_buffer, cur_line, killed);
     erase_substr(b, pos, killed);
@@ -241,7 +241,7 @@ void end_region(Buffer *b) {
 void kill_region(Buffer *b) {
     if (b->reg_begin == b->reg_end) return;
 
-    size_t reg_len = b->reg_end - b->reg_begin;
+    uint32_t reg_len = b->reg_end - b->reg_begin;
 
     strncat(kill_buffer, b->reg_begin, reg_len);
     erase_substr(b, b->reg_begin - b->data, reg_len);
@@ -256,15 +256,15 @@ void kill_region(Buffer *b) {
 void copy_region(Buffer *b) {
     if (b->reg_begin == b->reg_end) return;
 
-    size_t reg_len = b->reg_end - b->reg_begin;
+    uint32_t reg_len = b->reg_end - b->reg_begin;
 
     strncat(kill_buffer, b->reg_begin, reg_len);
 }
 
 void paste(Buffer *b) {
-    size_t len = strlen(kill_buffer);
+    uint32_t len = strlen(kill_buffer);
     size_t pos = get_current_pos(b);
-    size_t count = line_count(kill_buffer);
+    uint32_t count = line_count(kill_buffer);
 
     if (len == 0) return;
